feat: 주문할 때 메뉴와 수량 입력 추가

+++ src/miniproject/Main.java
@@ -31,26 +31,41 @@ public class Main {
                 
             // 메뉴 추가
             case 1:
-                System.out.print("추가할 메뉴의 이름을 입력하세요: ");
-                String menuName = sc.nextLine();
-                
-                System.out.print("추가할 메뉴의 가격을 입력하세요: ");
-                double menuPrice = sc.nextDouble();
-                sc.nextLine();
-                
-                restaurant.addMenu(menuName, menuPrice);
-                System.out.println(menuName + " 메뉴가 추가되었습니다.");
+            	while (true) {
+                    System.out.print("추가할 메뉴의 이름을 입력하세요: ");
+                    String menuName = sc.nextLine();
+                    
+                    if (restaurant.getMenu(menuName) != null) {
+                        System.out.println("이미 존재하는 메뉴 이름입니다. 다른 이름을 입력해주세요.");
+                    } else {
+                        System.out.print("추가할 메뉴의 가격을 입력하세요: ");
+                        double menuPrice = sc.nextDouble();
+                        sc.nextLine();
+                        
+                        restaurant.addMenu(menuName, menuPrice);
+                        System.out.println(menuName + " 메뉴가 추가되었습니다.");
+                        break;
+                    }
+                }
                 break;
 
             
             //메뉴 삭제 
             case 2:
-            	System.out.print("삭제할 메뉴의 이름을 입력하세요: ");
-                String deleteMenuName = sc.nextLine();
-                if (restaurant.deleteMenu(deleteMenuName)) {
-                    System.out.println(deleteMenuName + " 메뉴가 삭제되었습니다.");
-                } else {
-                    System.out.println("해당 메뉴가 존재하지 않습니다.");
+            	while (true) {
+                    System.out.print("삭제할 메뉴의 이름을 입력하세요: ");
+                    String deleteMenuName = sc.nextLine();
+                    
+                    if (restaurant.getMenu(deleteMenuName) == null) {
+                        System.out.println("존재하지 않는 메뉴 이름입니다. 다른 이름을 입력해주세요.");
+                    } else {
+                        if (restaurant.deleteMenu(deleteMenuName)) {
+                            System.out.println(deleteMenuName + " 메뉴가 삭제되었습니다.");
+                        } else {
+                            System.out.println("메뉴 삭제에 실패했습니다.");
+                        }
+                        break;
+                    }
                 }
                 break;
             
@@ -78,13 +93,21 @@ public class Main {
             
             // 테이블 삭제
             case 5:
-            	 System.out.print("삭제할 테이블 이름을 입력하세요: ");
-                 String removeTableName = sc.nextLine();
-                 if (restaurant.deleteTable(removeTableName)) {
-                     System.out.println(removeTableName + " 테이블이 삭제되었습니다.");
-                 } else {
-                     System.out.println("유효하지 않은 테이블입니다.");
-                 }
+            	while (true) {
+                    System.out.print("삭제할 테이블 이름을 입력하세요: ");
+                    String removeTableName = sc.nextLine();
+
+                    if (restaurant.getTable(removeTableName) == null) {
+                        System.out.println("존재하지 않는 테이블 번호입니다. 다른 번호를 입력해주세요.");
+                    } else {
+                        if (restaurant.deleteTable(removeTableName)) {
+                            System.out.println(removeTableName + " 테이블이 삭제되었습니다.");
+                        } else {
+                            System.out.println("테이블 삭제에 실패했습니다.");
+                        }
+                        break;
+                    }
+                }
                  break;
                  
               // 테이블 목록 출력
@@ -95,24 +118,22 @@ public class Main {
             
             // 테이블 번호 + 주문
             case 7:
-            	System.out.print("주문받은 테이블 이름을 입력하세요: ");
+            	
+            	System.out.print("주문받을 테이블 이름을 입력하세요: ");
                 String orderTableName = sc.nextLine();
                 Table orderTable = restaurant.getTable(orderTableName);
+
                 if (orderTable != null) {
                     System.out.print("주문할 메뉴 이름을 입력하세요: ");
                     String orderMenuName = sc.nextLine();
-
-                    Menu[] menuList = restaurant.getMenuList();
-                    Menu orderMenu = null;
-                    for (Menu menu : menuList) {
-                        if (menu != null && menu.getName().equals(orderMenuName)) {
-                            orderMenu = menu;
-                            break;
-                        }
-                    }
+    
+                    Menu orderMenu = restaurant.getMenu(orderMenuName);
                     if (orderMenu != null) {
-                        orderTable.addOrder(orderMenu);
-                        System.out.println(orderTable.getTableName() + " 테이블: " + orderMenu.getName() + " 주문 완료");
+                        System.out.print("주문할 수량을 입력하세요: ");
+                        int quantity = sc.nextInt();
+                        sc.nextLine();
+    
+                        orderTable.addOrder(orderMenu, quantity);
                     } else {
                         System.out.println("유효하지 않은 메뉴 이름입니다.");
                     }
